ZimWikiReplacePatternGeneratorTests
